Moved the classes back so their package wasn't nested inside a string of folders that was the same (so that the directory isnt edu/kennesaw/seclass/gradescalc/edu/kennesaw/seclass/gradescalc". made Student.getAttendance() return an int to pass testGetStudentsByName2

+++ edu/kennesaw/seclass/gradescalc/Student.java
@@ -27,7 +27,7 @@ public class Student {
 		return id;
 	}
 	
-	public double getAttendance() { return attendance;	}
+	public int getAttendance() { return (int)attendance;	}
 	
 	public void setName(String n)
 	{
